requirements.txt
Python==3.7.3
Django==3.2.2
djangorestframework==3.12.4